requirements.txt
opencv-python
numpy
pillow
matplotlib
pytesseract
scikit-image
imutils
pandas
tqdm
loguru
streamlit
